+++ ttnn/api/ttnn/tensor/tensor_utils.hpp
@@ -41,21 +41,6 @@ uint32_t get_batch_size(const T& shape) {
     return result;
 }
 
-// Useful information about how a shard_shape cuts a 2D shape
-// - num_shards_height: Number of shards along the height (including partial last shard, if any)
-// - last_shard_height: Height of last partial shard (if None, it will be same as full shard shape height)
-// - num_shards_width: Number of shards along the width (including partial last shard, if any)
-// - last_shard_width: Width of last partial shard (if None, it will be same as full shard shape width)
-struct ShardDivisionSpec {
-    size_t num_shards_height = 0;
-    size_t last_shard_height = 0;
-    size_t num_shards_width = 0;
-    size_t last_shard_width = 0;
-};
-
-// Returns ShardDivisionSpecs given 2D shape and shard_shape
-ShardDivisionSpec compute_shard_division_spec(const Shape2D& shape, const Shape2D& shard_shape);
-
 /**
  * @brief Creates a CBDescriptor from a sharded tensor.
  *
